# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/user/me": {
      "get": {
        "operationId": "UserController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns user details"
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get details of the logged-in user",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/all": {
      "get": {
        "operationId": "UserController_getAllUser",
        "parameters": [
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of users per page",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of users"
          },
          "403": {
            "description": "Forbidden for non-admin users"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users with optional filters",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/cart": {
      "post": {
        "operationId": "UserController_addToCart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item added to the cart successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an item to the cart",
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateCartItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart item updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an item in the cart",
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "UserController_getCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns cart details"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the items in the cart",
        "tags": [
          "User"
        ]
      }
    },
    "/api/v1/user/cart/{id}": {
      "delete": {
        "operationId": "UserController_removeItemFromCart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cart item removed successfully"
          },
          "404": {
            "description": "Cart item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an item from the cart",
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/v1/order": {
      "post": {
        "operationId": "OrderController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully"
          },
          "400": {
            "description": "Invalid request payload"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new order",
        "tags": [
          "Order"
        ]
      },
      "get": {
        "operationId": "OrderController_getOrders",
        "parameters": [
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderStatus",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "paymentStatus",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of orders per page",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of orders"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch orders with optional filters",
        "tags": [
          "Order"
        ]
      }
    },
    "/api/v1/order/status/{OrderId}": {
      "patch": {
        "operationId": "OrderController_updateOrderStatus",
        "parameters": [
          {
            "name": "OrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated successfully"
          },
          "403": {
            "description": "Forbidden for unauthorized roles"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the status of an order",
        "tags": [
          "Order"
        ]
      }
    },
    "/api/v1/order/{OrderId}": {
      "get": {
        "operationId": "OrderController_getOrdersByOrderId",
        "parameters": [
          {
            "name": "OrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns order details"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch order details by order ID",
        "tags": [
          "Order"
        ]
      }
    },
...
  },
  "info": {
    "title": "Book API Documentation",
    "description": "The Book API provides a comprehensive set of endpoints for managing books, categories, authors, and users in a book management system. This API supports the ability to perform CRUD operations on books, handle user authentication, and manage categories and authors. It is designed to be scalable, secure, and easy to integrate with various client applications. ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CartDto": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "bookId",
          "quantity"
        ]
      },
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "shippingId": {
            "type": "string"
          }
        },
        "required": [
          "shippingId"
        ]
      },
      "UpdateOrderStatus": {
        "type": "object",
        "properties": {
          "orderStatus": {
            "type": "string"
          }
        },
        "required": [
          "orderStatus"
        ]
      },
...
    }
  }
}